Teolemon's requests 1
Warning/Information inside the diets fragments.
Test that field name is not empty before saving a diet. Message to the user if field name is empty.
Cleaning code comments.
Rename a diet doesn't create a new diet anymore.
Not really satisfying correction of the crash due to unclosed parenthesis in ingredients list.
Suppress asterisk(s) in name of ingredient.