Correct "install/open" drawer item labels (#2547)
* Correct "install/open" drawer item labels

The drawer item to allow install or opening of variant apps now
correctly reflects which variant are currently installed on the
device. This change fixes #2533

* Add string "open_other_flavor_drawer" & simplify implementation

These changes greatly simplify the previous fix by introducing
the translatable string "Install" and concatenating it to the
new string "open_other_flavor_drawer" that will hold the value
of whichever app variant is to appear in the drawer to open or
install.