Change theme and fix bottom navigation issues (#2894)
* WIP: Fixed menu highlights for the main bottom navigation

The current fix uses setSelectedItemId to make sure that the correct
menu item is selected. This API was introduced in SDK 25.3.0.

* Changed the bottom navigation color scheme to see clearly what item is selected

* Removed redundant bottom navigation declaration

* Fixed active/inactive size + deselected the fist item for activities not mentioned in the bottom navigation bar

* Make sure that the Home navigation item is highlighted on activity resume

* Made the Comparison activity to redirect to Home like all activities

A problem with this was that onOptionItemSelected returned true by
default which blocked the call from completing. In our case, we don't
have to override the call.

Also, the manifest needed to be modified to redirect you to the home
activity.

* Reworked bottom navigation logic + added compatibility code

* Fixed trucated navigation item when the text was to large

Fixed the issue described into this material fix:
https://github.com/material-components/material-components-android/issues/139