Fix DB closed crash at startup
Crash was caused by the try with resource insatntiation on the DB helper. https://docs.oracle.com/javase/tutorial/essential/exceptions/tryResourceClose.html says that an object implementing Closeable created in a try with resource instation will be closed at the end of the try block, and our DatabaseHelper does implements AutoCloseable, so it is closed right after begin created, therefore the db is closed and never accessible from the app.