Fix dark mode error in Contributions tab (#4628)
`TextView` In fragment_contributors.xml the TextView for "Product
States" has text color set to black, which kept the text black when in
dark mode. After fix, the text color is light when in dark mode, and
dark when dark in light mode.

I used `textAppearance` because it is used in `attribute_group_item.xml`

Co-authored-by: galemu00 <galemu00@galemu00-Nitro-AN515-51>